Plot point values on a map with a colour function.

Add some coloured markers for observations on an existing map.

Usage

obsplot(x,y,z,breaks=5,pretty=TRUE,legend.pos=NULL,
        add=TRUE,domain=.Last.domain(),col=irainbow,...)

Arguments

x,y

Vectors with co-ordinates. If y is missing, x is suposed to have two columns x and y.

z

Vector with data (same length as x and y).

breaks

The number of breaks in the colour code.

pretty

If TRUE, the breaks are chosen at 'nice' values.

legend.pos

Legend position.

add

If TRUE, add to current plot. Otherwise, start a new plot with dmain.

domain

By default, the last plotted domain.

col

Colour scheme.

...

Options for 'points'.

Value

A set of points is plotted at the x,y points, with colour taken from the chosen colour scheme. The return value (hidden) is a list wit hthe projected x,y points, the data values, levels and colours.

See also