Plotting a map for a given domain.

This function plots country borders for a given domain.

Usage

# S3 method for geodomain
plot(x=.Last.domain(), add=FALSE,
              col=1, mapfill=c("sandybrown","steelblue"),
              add.dx=TRUE, box=TRUE,
              fill=FALSE, interior=TRUE,
              map.database="world", asp=1, ...)
# S3 method for geofield
plot(x, ...)

Arguments

x

A an object of geofield or geodomain class. Default is the last domain plotted.

col

Colour of the map contours.

mapfill

Colour(s) used if fill=TRUE to fill the map contours. First colour is for internal (land), second for background (sea), but this only happens if add=FALSE).

fill

If TRUE, the map is drawn as coloured polygons, not just lines.

interior

If FALSE, only the physical land boundaries are shown, not the interior political borders.

map.database

A character string for the map database. Default is 'world' from the maps package.

add.dx

If TRUE, the domain size is extended by half a grid interval in all directions, i.e. the co-ordinates are treated as the centers of grid boxes.

box

If TRUE, a rectangular box is drawn around the domain.

add

If TRUE, the plot is superposed on the current plot.

asp

Aspect ratio. Not considered if add=TRUE. For Lat/Lon domains, NULL may be a better choice!.

...

Other standard graphical parameters.

Details

These plot methods only plot a map. For geofield objects there are various other functions that plot the actual data.

See also

Examples

if (FALSE) {
ttt <- FAdec(...)
### to plot the domain of ttt (not the data itself):
plot(ttt, mapreso=.01, add=FALSE)
}